requirements.txt
chainlit==1.0.500
datasets==2.15.0
fitz==0.0.1.dev2
jsonlines==4.0.0
openai==0.28.0
pandas==2.2.1
pydantic==2.6.4
pyprojroot==0.3.0
python-dotenv==1.0.1
PyYAML==6.0.1
torch==2.1.1
transformers==4.37.0
